# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Game World Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/status:
    patch:
      operationId: updateMyStatus
      tags: [auth]
      summary: Update online status and current game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /auth/login/2fa:
    post:
      operationId: verifyTwoFactorLogin
      tags: [auth]
      summary: Complete a two-factor login challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired challenge/code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password-reset/request:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset code by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequestInput"
      responses:
        "200":
          description: Always ok (no account enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags: [auth]
      summary: Set a new password using an emailed reset code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmInput"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/email:
    post:
      operationId: setMyEmail
      tags: [auth]
      summary: Set or change the recovery email (sends a verification code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetEmailInput"
      responses:
        "200":
          description: Verification code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/email/verify:
    post:
      operationId: verifyMyEmail
      tags: [auth]
      summary: Verify the recovery email with the emailed code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/2fa/setup:
    post:
      operationId: setupTwoFactor
      tags: [auth]
      summary: Begin enabling two-factor auth (returns TOTP secret or emails a code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorSetupInput"
      responses:
        "200":
          description: Setup info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorSetupResponse"
        "400":
          description: Precondition failed (e.g. email not verified)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/2fa/enable:
    post:
      operationId: enableTwoFactor
      tags: [auth]
      summary: Confirm and enable two-factor auth with a code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorEnableInput"
      responses:
        "200":
          description: Two-factor enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/2fa/disable:
    post:
      operationId: disableTwoFactor
      tags: [auth]
      summary: Disable two-factor auth (requires current password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorDisableInput"
      responses:
        "200":
          description: Two-factor disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Users ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # ── LFG (Looking For Group) ───────────────────────────────────────────────
...
    RegisterInput:
      type: object
      required: [username, password, displayName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 6
        displayName:
          type: string
          minLength: 1
          maxLength: 50
        email:
          type: string
          maxLength: 255
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        requiresTwoFactor:
          type: boolean
        twoFactorMethod:
          type: string
          enum: [email, totp]
        challengeToken:
          type: string
    StatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [online, away, busy, offline]
        currentGame:
          type: ["string", "null"]

    # ── Users ────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, username, displayName, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        twoFactorMethod:
          type: string
          enum: [none, email, totp]
        allowProfileComments:
          type: boolean
        status:
          type: string
          enum: [online, away, busy, offline]
        currentGame:
          type: ["string", "null"]
        createdAt:
          type: string
...
    SetEmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          maxLength: 255
    VerifyEmailInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
    PasswordResetRequestInput:
      type: object
      required: [identifier]
      properties:
        identifier:
          type: string
          description: Username or email address
          minLength: 1
    PasswordResetConfirmInput:
      type: object
      required: [identifier, code, newPassword]
      properties:
        identifier:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 6
          maxLength: 6
        newPassword:
          type: string
          minLength: 6
    TwoFactorLoginInput:
      type: object
      required: [challengeToken, code]
      properties:
        challengeToken:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
    TwoFactorSetupInput:
      type: object
      required: [method]
      properties:
        method:
          type: string
          enum: [email, totp]
    TwoFactorSetupResponse:
      type: object
      required: [method]
      properties:
        method:
          type: string
          enum: [email, totp]
        secret:
          type: string
          description: TOTP secret (base32) — only for method totp
        otpauthUrl:
          type: string
          description: otpauth:// URL to render as QR — only for method totp
    TwoFactorEnableInput:
      type: object
      required: [method, code]
      properties:
        method:
          type: string
          enum: [email, totp]
        code:
          type: string
          minLength: 6
          maxLength: 6
    TwoFactorDisableInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
...