# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Game World Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  /** @minLength 6 */
  password: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  displayName: string;
  /** @maxLength 255 */
  email?: string;
}