# lib/api-zod/src/generated/types/setEmailInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Game World Hub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SetEmailInput {
  /** @maxLength 255 */
  email: string;
}