# lib/api-zod/src/generated/types/twoFactorEnableInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Game World Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TwoFactorEnableInputMethod } from './twoFactorEnableInputMethod';

export interface TwoFactorEnableInput {
  method: TwoFactorEnableInputMethod;
  /**
     * @minLength 6
     * @maxLength 6
     */
  code: string;
}