# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Game World Hub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from './userStatus';
import type { UserTwoFactorMethod } from './userTwoFactorMethod';

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  email?: string | null;
  emailVerified?: boolean;
  twoFactorMethod?: UserTwoFactorMethod;
  allowProfileComments?: boolean;
  status: UserStatus;
  /** @nullable */
  currentGame?: string | null;
  createdAt: string;
}